Changed such improvements as infinite mana and infinite water. Due to the fact that it was possible to extract mana/water from the mechanism, the logic of these improvements was changed. Now they remove the consumption of mana/water in the recipe.
The process of crafting in the Rune Altar has been changed. Now the craft consumes as much livingrock as would be needed for 1 recipe (instead of 1 livingrock for several recipes). Therefore, a certain buffer has appeared that stores the amount of livingrock that needs to be given to the mechanism if the recipe consumed more than it had at the time the recipe was completed. Also in the apothecary (Seeds in it).